# api-gateway/docs/swagger.yaml
basePath: /
...
info:
  contact: {}
  description: |-
    Gateway for Mallow algorithmic trading system.

    All protected routes require a Bearer JWT token obtained from the Identity service (`/api/v1/auth/login`).
  title: Mallow API Gateway
  version: "1.0"
paths:
...
  /api/v1/hands:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: List hands
      tags:
      - Helm
    post:
      consumes:
      - application/json
      description: Creates a new signal-following bot under a helm. The hand starts
        stopped; call /start to activate.
      parameters:
      - description: Hand config
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a hand (autonomous bot)
      tags:
      - Helm
  /api/v1/hands/{id}:
    delete:
      parameters:
      - description: Hand ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete a hand
      tags:
      - Helm
    get:
      parameters:
      - description: Hand ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a hand
      tags:
      - Helm
    put:
      consumes:
      - application/json
      parameters:
      - description: Hand ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated config
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update hand config
      tags:
      - Helm
  /api/v1/hands/{id}/kill:
    post:
      description: Stops the hand and flattens its positions at the exchange.
      parameters:
      - description: Hand ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Kill a hand (flatten positions)
      tags:
      - Helm
  /api/v1/hands/{id}/pause:
    post:
      parameters:
      - description: Hand ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Pause a hand
      tags:
      - Helm
...
  /api/v1/hands/{id}/resume:
    post:
      parameters:
      - description: Hand ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Resume a paused hand
      tags:
      - Helm
...
  /api/v1/helms:
    get:
      description: Returns all Helm accounts accessible to the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: List helms
      tags:
      - Helm
  /api/v1/helms/{id}:
    get:
      parameters:
      - description: Helm ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a helm
      tags:
      - Helm
    put:
      consumes:
      - application/json
      parameters:
      - description: Helm ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated config
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update helm config
      tags:
      - Helm
  /api/v1/helms/{id}/disable:
    post:
      parameters:
      - description: Helm ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Disable a helm
      tags:
      - Helm
  /api/v1/helms/{id}/enable:
    post:
      parameters:
      - description: Helm ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Enable a helm
      tags:
      - Helm
  /api/v1/helms/{id}/halt/reset:
    post:
      description: Clears the halted state so the helm can be re-enabled.
      parameters:
      - description: Helm ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Reset a halted helm
      tags:
      - Helm
  /api/v1/helms/{id}/kill:
    post:
      description: Stops all hands and flattens their positions at the exchange. Sets
        helm state to halted.
      parameters:
      - description: Helm ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Kill a helm (flatten all positions)
      tags:
      - Helm
  /api/v1/helms/{id}/orders:
    get:
      parameters:
      - description: Helm ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Helm open orders
      tags:
      - Helm
  /api/v1/helms/{id}/pause:
    post:
      parameters:
      - description: Helm ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Pause a helm (cascade-stops all hands)
      tags:
      - Helm
  /api/v1/helms/{id}/portfolio:
    get:
      parameters:
      - description: Helm ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Helm portfolio snapshot
      tags:
      - Helm
  /api/v1/helms/{id}/positions:
    get:
      parameters:
      - description: Helm ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Helm open positions
      tags:
      - Helm
  /api/v1/helms/{id}/resume:
    post:
      parameters:
      - description: Helm ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Resume a paused helm
      tags:
      - Helm
  /api/v1/helms/{id}/trades:
    get:
      parameters:
      - description: Helm ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Helm trade history
      tags:
      - Helm
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: JWT token — prefix with "Bearer "
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"